Add unit tests for pull-queue.ts
This adds unit tests or pull-queue.ts. This also uses classes and sets
from ES2015 and up to update this a little.

It was a little easier to refactor this knowing I could cross check
the new implementation against the unit tests that worked on the
previous iteration of PullQueue.ts.